pyproject.toml
[project]
name = "airline-data-system"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    #    "azure-ai-textanalytics>=5.3.0",
    #    "azure-identity>=1.23.0",
    #    "azureml-sdk>=1.0.23",
    "bandit>=1.8.6",
    "beautifulsoup4>=4.13.4",
    "black>=25.1.0",
    "detect-secrets>=1.5.0",
    "dotenv-linter>=0.7.0",
    "folium>=0.20.0",
    "geopy>=2.4.1",
    "httpx>=0.28.1",
    "ipython>=9.4.0",
    "isort>=6.0.1",
    "jupyterlab>=4.4.4",
    "lxml>=6.0.0",
    "mypy>=1.16.1",
    "numpy>=2.3.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "pre-commit>=4.2.0",
    "psycopg2-binary>=2.9.10",
    "pyarrow>=20.0.0",
    "pylint>=3.3.7",
    "pymongo>=4.13.2",
    "pyspark>=4.0.0",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "ruff>=0.12.3",
    "safety>=3.6.0",
    "scikit-learn>=1.7.0",
    "sqlalchemy>=2.0.41",
    "streamlit>=1.46.1",
    "taskipy>=1.14.1",
    "tox>=4.27.0",
    "xgboost>=3.0.2",
    "yamllint>=1.37.1",
    "mongomock>=4.0.0",
    "autoflake>=2.3.1",
]
